requirements.txt
crewai
crewai[tools]
arize-phoenix
opentelemetry-sdk
opentelemetry-exporter-otlp
openinference-instrumentation-crewai
